+++ packages/brick_offline_first_with_rest_build/test/offline_first_generator/test_default_value.dart
@@ -31,7 +31,7 @@ class DefaultValue {
     this.string,
   });
 
-  @Rest(defaultValue: 'Thomas')
-  @Sqlite(defaultValue: 'Guy')
+  @Rest(defaultValue: '"Thomas"')
+  @Sqlite(defaultValue: '"Guy"')
   final String string;
 }