+++ packages/brick_rest/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* `Rest#defaultValue` updated to reflect `FieldSerializable#defaultValue` change
+
 ## 0.0.2
 
 * Fix linter hints