+++ packages/brick_sqlite_abstract/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* `Sqlite#defaultValue` updated to reflect `FieldSerializable#defaultValue` change
+
 ## 0.0.2
 
 * Use single quotes when generating `Migration#version` to comply with default analyzer options